secboot: address review comment
